test: fix batching test suite in browsers
