+++ packages/platform-test-suite/package.json
@@ -44,6 +44,7 @@
     "dirty-chai": "^2.0.1",
     "dotenv-safe": "^8.2.0",
     "events": "^3.3.0",
+    "glob": "^10.3.4",
     "https-browserify": "^1.0.0",
     "js-merkle": "^0.1.5",
     "karma": "^6.4.1",
