+++ packages/platform-test-suite/test/functional/platform/proofs.spec.js
@@ -24,6 +24,8 @@ function executeProof() {
 
 }
 
+// TODO: Fix test to be running in Browser
+
 describe('Platform', () => {
   describe('Proofs', () => {
     let blake3;
